Updating validation procedures.
