+++ src/api/list/listPost.js
@@ -2,10 +2,11 @@ const express = require('express')
 
 const router = express.Router()
 
+const { validationObjectSize, validateParams } = require('../../validation')
 const { listDataStore } = require('../../data')
 
 /*
-{
+const body = {
 	"name":"Erica",
 	"email": "a@aol.com",
 	"phone": "562",
@@ -14,17 +15,8 @@ const { listDataStore } = require('../../data')
 }
 */
 
-const validation = body => {
-	const keys = Object.keys(body)
-	console.log(keys, keys.length)
-	if (keys.length === 5) {
-		return keys.length // true. 5, '5', {}?, []?
-	}
-	return 0 // false, 0, '0', null, undefined
-}
-
 router.post('/', (req, res) => {
-	if (validation(req.body)) {
+	if (validationObjectSize(req.body) && validateParams(req.body)) {
 		console.log('success: ', req.body)
 		listDataStore.push(req.body)
 		return res.json('POST SUCCESSFUL!!! And matt is a bitch!')
